+++ app/javascript/vue/tasks/collecting_events/new_collecting_event/app.vue
@@ -42,57 +42,65 @@
             <radial-object :global-id="collectingEvent.global_id"/>
           </template>
         </div>
-        <div class="horizontal-right-content">
-          <span class="margin-small-right">Collection object:</span>
-          <button
-            type="button"
-            class="button normal-input button-default margin-small-right"
-            @click="openComprehensive"
-            :disabled="!collectingEvent.id">
-            New
-          </button>
-          <collection-objects-table
-            class="margin-small-right"
-            :ce-id="collectingEvent.id"/>
-          <parse-data
-            class="margin-large-right"
-            @onParse="setCollectingEvent"/>
-          <button
-            @click="showRecent = true"
-            class="button normal-input button-default button-size margin-small-right"
-            type="button">
-            Recent
-          </button>
-          <navigate-component
-            class="margin-small-right"
-            :collectingEvent="collectingEvent"
-            @select="loadCollectingEvent"
-          />
-          <button
-            type="button"
-            class="button normal-input button-submit margin-small-right"
-            :disabled="!collectingEvent.id"
-            @click="cloneCE"
-          >
-            Clone
-          </button>
-          <button
-            v-shortkey="[getOSKey(), 's']"
-            @shortkey="saveCollectingEvent"
-            @click="saveCollectingEvent"
-            class="button normal-input button-submit button-size margin-small-right"
-            type="button">
-            Save
-          </button>
-          <button
-            v-shortkey="[getOSKey(), 'n']"
-            @shortkey="reset"
-            @click="reset"
-            class="button normal-input button-default button-size"
-            type="button">
-            New
-          </button>
-        </div>
+        <ul class="context-menu no_bullets">
+          <li class="horizontal-right-content">
+            <span class="margin-small-right">Collection object:</span>
+            <button
+              type="button"
+              class="button normal-input button-default margin-small-right"
+              @click="openComprehensive"
+              :disabled="!collectingEvent.id">
+              New
+            </button>
+            <collection-objects-table
+              class="margin-small-right"
+              :ce-id="collectingEvent.id"/>
+            <parse-data
+              @onParse="setCollectingEvent"/>
+          </li>
+          <li class="horizontal-right-content">
+            <span
+              v-if="isUnsaved"
+              class="medium-icon margin-small-right"
+              title="You have unsaved changes."
+              data-icon="warning"/>
+            <button
+              @click="showRecent = true"
+              class="button normal-input button-default button-size margin-small-right"
+              type="button">
+              Recent
+            </button>
+            <navigate-component
+              class="margin-small-right"
+              :collectingEvent="collectingEvent"
+              @select="loadCollectingEvent"
+            />
+            <button
+              type="button"
+              class="button normal-input button-submit margin-small-right"
+              :disabled="!collectingEvent.id"
+              @click="cloneCE"
+            >
+              Clone
+            </button>
+            <button
+              v-shortkey="[getOSKey(), 's']"
+              @shortkey="saveCollectingEvent"
+              @click="saveCollectingEvent"
+              class="button normal-input button-submit button-size margin-small-right"
+              type="button">
+              Save
+            </button>
+            <button
+              v-shortkey="[getOSKey(), 'n']"
+              @shortkey="reset"
+              @click="reset"
+              class="button normal-input button-default button-size"
+              type="button">
+              New
+            </button>
+          </li>
+        </ul>
       </div>
     </nav-bar>
     <recent-component
@@ -132,25 +140,16 @@ import NavBar from 'components/navBar'
 import ParseData from './components/parseData'
 
 import CollectingEventForm from './components/CollectingEventForm'
-import makeCollectingEvent from './const/collectingEvent'
 import CollectionObjectsTable from './components/CollectionObjectsTable.vue'
 import NavigateComponent from './components/Navigate'
 import SpinnerComponent from 'components/spinner'
 
-import {
-  CloneCollectionEvent,
-  CreateCollectingEvent,
-  CreateIdentifier,
-  CreateLabel,
-  GetCollectingEvent,
-  GetCollectionObject,
-  GetLabelsFromCE,
-  GetSoftValidation,
-  GetTripCodeByCE,
-  UpdateCollectingEvent,
-  UpdateIdentifier,
-  UpdateLabel
-} from './request/resources'
+import { ActionNames } from './store/actions/actions'
+import { GetterNames } from './store/getters/getters'
+import { MutationNames } from './store/mutations/mutations'
+
+import { GetCollectionObject } from './request/resources'
+import updateLastChange from './store/mutations/updateLastChange'
 
 export default {
   components: {
@@ -170,23 +169,37 @@ export default {
   computed: {
     collectingEvent: {
       get () {
-        return this.ce
+        return this.$store.getters[GetterNames.GetCollectingEvent]
       },
       set (value) {
-        this.ce = value
+        this.$store.commit(MutationNames.SetCollectingEvent, value)
       }
+    },
+    isUnsaved () {
+      return this.$store.getters[GetterNames.IsUnsaved]
+    },
+    isLoading () {
+      return this.$store.getters[GetterNames.GetSettings].isLoading
+    },
+    isSaving () {
+      return this.$store.getters[GetterNames.GetSettings].isSaving
     }
   },
   data () {
     return {
-      isLoading: false,
-      isSaving: false,
       settings: {
         sortable: false
       },
       showRecent: false,
-      validation: [],
-      ce: makeCollectingEvent()
+      validation: []
+    }
+  },
+  watch: {
+    collectingEvent: {
+      handler (newVal) {
+        this.$store.commit(MutationNames.UpdateLastChange)
+      },
+      deep: true
     }
   },
   mounted () {
@@ -210,94 +223,25 @@ export default {
   },
   methods: {
     cloneCE () {
-      CloneCollectionEvent(this.collectingEvent.id).then(response => {
-        this.setCollectingEvent(response.body)
-        TW.workbench.alert.create('Collecting event was successfully cloned.', 'notice')
-      })
+      this.$store.dispatch(ActionNames.CloneCollectingEvent)
     },
     reset () {
-      this.ce = makeCollectingEvent()
-      this.validation = []
+      this.$store.dispatch(ActionNames.ResetStore)
       SetParam(RouteNames.NewCollectingEvent, 'collecting_event_id')
       SetParam(RouteNames.NewCollectingEvent, 'collection_object_id')
     },
     loadCollectingEvent (id) {
-      this.isLoading = true
-      GetCollectingEvent(id).then(async response => {
-        const ce = response.body
-        this.loadValidation(ce.global_id)
-        const label = (await GetLabelsFromCE(ce.id)).body[0]
-        const tripCode = (await GetTripCodeByCE(ce.id)).body[0]
-        ce.label = label || makeCollectingEvent().label
-        ce.tripCode = tripCode || makeCollectingEvent().tripCode
-        ce.roles_attributes = ce.collector_roles || []
-        this.setCollectingEvent(response.body)
-      }).finally(() => {
-        this.isLoading = false
-      })
+      this.$store.dispatch(ActionNames.LoadCollectingEvent, id)
     },
     setCollectingEvent (ce) {
       this.collectingEvent = Object.assign({}, this.collectingEvent, ce)
       SetParam(RouteNames.NewCollectingEvent, 'collecting_event_id', this.collectingEvent.id)
     },
-    async saveCollectingEvent () {
-      const saveCE = this.collectingEvent.id ? UpdateCollectingEvent : CreateCollectingEvent
-      const cloneCE = this.cleanCE()
-
-      if (this.collectingEvent.units === 'ft') {
-        ['minimum_elevation', 'maximum_elevation', 'elevation_precision'].forEach(key => {
-          const elevationValue = Number(cloneCE[key])
-          cloneCE[key] = elevationValue > 0 ? elevationValue / 3.281 : undefined
-        })
-        this.collectingEvent.units = undefined
-      }
-
-      this.isSaving = true
-      saveCE(cloneCE).then(async response => {
-        this.loadValidation(response.body.global_id)
-        response.body.label = await this.saveLabel(this.collectingEvent)
-        response.body.tripCode = await this.saveIdentifier(this.collectingEvent.tripCode)
-        TW.workbench.alert.create(`Collecting event was successfully ${this.collectingEvent.id ? 'updated' : 'created'}.`, 'notice')
-        this.setCollectingEvent(response.body)
-      }).finally(() => {
-        this.isSaving = false
-      })
-    },
-    async saveLabel (ce) {
-      const label = this.collectingEvent.label
-
-      if (label.text.length && label.total) {
-        label.label_object_id = ce.id
-        return label.id ? (await UpdateLabel(label.id, { label: label })).body : (await CreateLabel({ label: label })).body
-      } else {
-        return label
-      }
-    },
-    async saveIdentifier (identifier) {
-      if (identifier.namespace_id && identifier.identifier) {
-        identifier.identifier_object_id = this.collectingEvent.id
-        return identifier.id ? (await UpdateIdentifier(identifier)).body : (await CreateIdentifier(identifier)).body
-      } else {
-        return identifier
-      }
-    },
-    loadValidation (globalId) {
-      GetSoftValidation(globalId).then(response => {
-        this.validation = response.body.validations.soft_validations
-      })
+    saveCollectingEvent () {
+      this.$store.dispatch(ActionNames.SaveCollectingEvent)
     },
     openComprehensive () {
       window.open(`${RouteNames.DigitizeTask}?collecting_event_id=${this.collectingEvent.id}`, '_self')
-    },
-    cleanCE () {
-      const cloneCE = JSON.parse(JSON.stringify(this.collectingEvent))
-      delete cloneCE.queueGeoreferences
-      delete cloneCE.label
-      delete cloneCE.geographicArea
-      delete cloneCE.georeferences
-      delete cloneCE.tripCode
-      delete cloneCE.units
-      return cloneCE
     },
     getOSKey: GetOSKey
   }