+++ app/javascript/vue/tasks/collecting_events/new_collecting_event/components/parsed/GeographicArea.vue
@@ -58,13 +58,13 @@
         </div>
       </modal-component>
     </template>
-    <template v-if="selected">
+    <template v-if="geographicArea">
       <div class="middle separate-top">
         <span data-icon="ok" />
-        <span class="separate-right"> {{ (selected['label'] ? selected.label : selected.name) }}</span>
+        <span class="separate-right"> {{ geographicArea.name }}</span>
         <span
           class="circle-button button-default btn-undo"
-          @click="clearSelection"
+          @click="selectGeographicArea()"
         />
       </div>
     </template>
@@ -73,9 +73,11 @@
 
 <script>
 
+import { GetterNames } from '../../store/getters/getters'
+import { ActionNames } from '../../store/actions/actions'
 import SmartSelector from 'components/smartSelector.vue'
-import { GetGeographicAreaByCoords, GetGeographicArea } from '../../request/resources.js'
-
+import { GetGeographicAreaByCoords } from '../../request/resources.js'
+import convertDMS from 'helpers/parseDMS.js'
 import ModalComponent from 'components/modal'
 
 import extendCE from '../mixins/extendCE'
@@ -89,6 +91,15 @@ export default {
   computed: {
     geographicAreaId () {
       return this.collectingEvent.geographic_area_id
+    },
+    verbatimLatitude () {
+      return this.collectingEvent.verbatim_latitude
+    },
+    verbatimLongitude () {
+      return this.collectingEvent.verbatim_longitude
+    },
+    geographicArea () {
+      return this.$store.getters[GetterNames.GetGeographicArea]
     }
   },
   data () {
@@ -97,35 +108,36 @@ export default {
       showModal: false,
       areasByCoors: [],
       geoId: undefined,
-      geographicArea: undefined
+      ajaxCall: undefined,
+      delay: 1000
     }
   },
   watch: {
     geographicAreaId (newVal) {
-      if (newVal) {
-        GetGeographicArea(newVal).then(response => {
-          this.collectingEvent.geographicArea = response.body
-          this.selected = response.body
-        })
-      } else {
-        this.clearSelection()
-      }
+      this.$store.dispatch(ActionNames.LoadGeographicArea, newVal)
+    },
+    verbatimLongitude () {
+      this.getGeographicByVerbatim()
+    },
+    verbatimLatitude () {
+      this.getGeographicByVerbatim()
     }
   },
   methods: {
-    clearSelection () {
-      this.selected = undefined
-      this.collectingEvent.geographic_area_id = null
-      this.collectingEvent.geographicArea = undefined
-    },
     selectGeographicArea (item) {
-      this.selected = item
-      this.collectingEvent.geographic_area_id = item.id
+      this.$store.dispatch(ActionNames.LoadGeographicArea, item?.id)
     },
     getByCoords (lat, long) {
       GetGeographicAreaByCoords(lat, long).then(response => {
         this.areasByCoors = response.body
       })
+    },
+    getGeographicByVerbatim () {
+      if (this.collectingEvent.geographic_area_id) return
+      if (convertDMS(this.verbatimLatitude) && convertDMS(this.verbatimLongitude)) {
+        clearTimeout(this.ajaxCall)
+        this.ajaxCall = setTimeout(() => { this.getByCoords(convertDMS(this.verbatimLatitude), convertDMS(this.verbatimLongitude)) }, this.delay)
+      }
     }
   }
 }