+++ app/javascript/vue/tasks/collecting_events/new_collecting_event/store/actions/loadSoftValidations.js
@@ -1,0 +1,8 @@
+import { GetSoftValidation } from '../../request/resources'
+import { MutationNames } from '../mutations/mutations'
+
+export default ({ commit }, globalId) => {
+  GetSoftValidation(globalId).then(response => {
+    commit(MutationNames.SetSoftValidations, response.body.validations.soft_validations)
+  })
+}